refactor code to separate components from pages
