+++ components/NftCard/index.tsx
