+++ components/Wrapper/index.tsx
@@ -1,10 +1,17 @@
 import React, { useState, useEffect } from 'react';
 import { useRouter } from 'next/router';
 import { GameContext } from '../../src/utils/web3';
+import { initInfo, useAppDispatch } from '../../src/redux';
 
 const Layout: React.FC = ({ children }) => {
   const router = useRouter();
+  const dispatch = useAppDispatch();
 
+  useEffect(() => {
+    dispatch(initInfo());
+  }, [dispatch]);
+
+  // TODO: Remove legacy code
   const [state, setState] = useState({
     web3: null,
     contract: null,
@@ -18,7 +25,7 @@ const Layout: React.FC = ({ children }) => {
   useEffect(() => {
     if (!state.loaded && router.asPath !== '/') {
       alert('wallet not connected');
-      router.push('/');
+      // router.push('/');
     }
   }, [state.loaded, router]);
