+++ src/redux/index.ts
@@ -1,0 +1,2 @@
+export * from './rootReducer';
+export * from './store';
